refactor(linter): move lsp relevant code into own file (#13243)
Moving all code behind `#[cfg(feature = "language_server")]` into own file.
